.gitreview: Remove port, and add defaultremote
The main reason for the change is to add a defaultremote, so that
the default remote used by "git review" is "origin", which is
the default when cloning a repository. That way, someone who
clones this repository, makes a change and submits it for review
doesn't get the following behavior from "git review"...

    Creating a git remote called 'gerrit' that maps to:
	ssh://<username>@git.adacore.com:<port>/gnatcoll-bindings

... indicating that "git review" is unnecessarily creating
a remote that basically duplicates the existing one.

Change-Id: Ida6925c8891bce2692beb6d740f5e97b631748b3