GMP: add bindings for rational arithmetic
Operators "+", "-" (both binary and unary), "*", "/", and "abs" have
been implemented.

Since operands have to be in a canonical form before any arithmetic
operations, a new parameter has been added to the `Set` directives to
handle that automatically. All rational values are canonicalized by
default but the user can still decide to not canonicalize a number if
needed. An exception is raised if an arithmetic operation with
non-canonical operands is performed.

TN: V310-009
Change-Id: If6dac7ed93f1e3042a3db4bdc45a77530664c47f